+++ pgvector/sqlalchemy/__init__.py
@@ -1,6 +1,7 @@
+from .bit import Bit
 from .halfvec import Halfvec
 from .sparsevec import Sparsevec
 from .vector import Vector
 from ..utils import SparseVec
 
-__all__ = ['Vector', 'Halfvec', 'Sparsevec']
+__all__ = ['Vector', 'Halfvec', 'Bit', 'Sparsevec']