+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 ## 0.3.0 (unreleased)
 
 - Added support for `halfvec`, `bit`, and `sparsevec` types to Django
-- Added support for `halfvec` and `sparsevec` types to SQLAlchemy and SQLModel
+- Added support for `halfvec`, `bit`, and `sparsevec` types to SQLAlchemy and SQLModel
 - Added support for `halfvec` and `sparsevec` types to Psycopg 3
 - Added support for `halfvec` and `sparsevec` types to Psycopg 2
 - Added support for `halfvec` and `sparsevec` types to asyncpg
 - Added support for `halfvec` and `sparsevec` types to Peewee
 - Added `L1Distance`, `HammingDistance`, and `JaccardDistance` for Django
-- Added `l1_distance` for SQLAlchemy, SQLModel, and Peewee
+- Added `l1_distance`, `hamming_distance`, and `jaccard_distance` for SQLAlchemy and SQLModel
+- Added `l1_distance` for Peewee
 
 ## 0.2.5 (2024-02-07)